Create script model & allow multiple script dirs
Big one

Overhaul interaction with scripts by placing all relevant methods in new
script class.
Contains 3 class methods for interaction with scripts as a
whole (gathering them from directories & sorting them).
Initialiser is weird as it needed to function off either the scripts
path & its name so it takes a hash with either a 'path' or 'name' key.
This process is in theory imperfect as determining a script's path from
its name or vis-a-versa could be ambiguous in systems with multiple
commands with the same name or nested script directories defined. As it
stands the party line is just to not do this.

Additionally allows >=1 directory for scripts to be defined in the
config file.
The order of these is important - the higher up directories are
'prioritised' and in the case of ambiguity the scripts in the higher
directories will be taken.
If '{nodename} is foudn in any of the script directory definitions it
will be replaced with the node's hostname

Fixes #7