Add 'chains'
Add -c (chain) option
Takes a path to a yaml file - each script & stage in which file is
executed in order. The file must be a valid yaml list of stages and
paths & stages are specified as 'stage: STAGE'
Additionally a role can be passed with the usual '-r' option - this will
make all stages then executed be executed for the specified node role.

Also move 'find_scripts_with_role_and_stage' method to Script class